Extract OMIS field definition out of create app
This places the OMIS fields at a higher level than just the create
app and journey so that it can be more easily shared between other
apps for example, edit.
